added select for changing currencies
